agents/requirements.txt
# ASI Alliance - uAgents Framework
uagents>=0.12.0
uagents-ai-engine>=0.4.0

# Web3 Integration
web3>=6.0.0
eth-account>=0.10.0
python-dotenv>=1.0.0

# API & Communication
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0
